add: logout confirm prompt

+++ app/Controllers/Login.php
@@ -41,7 +41,11 @@ class Login extends BaseController
 
     public function logout()
     {
-        $this->session->destroy();
-        return redirect()->to('/');
+        if ($this->request->getPost()) {
+            $this->session->destroy();
+            return redirect()->to('/');
+        }
+
+        echo view('logout');
     }
 }
